// web/lib/api/generated/model/freeGoodsScheme.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { AudienceTypeEnum } from './audienceTypeEnum';
import type { ProductRef } from './productRef';
import type { Ref } from './ref';
import type { ShopRef } from './shopRef';

export interface FreeGoodsScheme {
  readonly id: string;
  readonly name: string;
  readonly buy_product: ProductRef;
  readonly buy_unit: string;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  readonly buy_qty: string;
  readonly free_product: ProductRef;
  readonly free_unit: string;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  readonly free_qty: string;
  readonly repeat: boolean;
  /**
     * @nullable
     * @pattern ^-?\d{0,11}(?:\.\d{0,3})?$
     */
  readonly max_free_qty: string | null;
  readonly audience_type: AudienceTypeEnum;
  readonly price_list: Ref | null;
  readonly retailer: ShopRef | null;
  /** @nullable */
  readonly valid_from: string | null;
  /** @nullable */
  readonly valid_to: string | null;
  readonly is_active: boolean;
  readonly created_at: string;
  readonly updated_at: string;
}